Themes: Only reapply fonts, icons and overlays
The provider and WP service can be out of sync. This is possibly by applying
a theme and then separately going in and changing the WP.

Prior to this patch, an updated theme would trigger all set theme elements
to be reapplied, including the wallpaper of the set theme, even if the user
had changed the WP to something else.

The issue has been resolved by only applying theme elements which:
(1) do not pose a risk of being out of sync in the provider
(2) must be reapplied so as to not cause a FC.

Change-Id: I00f12952a5ef040e8f368d0fade222a3e88d003e

Conflicts:
	src/org/cyanogenmod/themes/provider/ThemePackageHelper.java